requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.1
amqp==5.2.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.2
billiard==4.2.1
celery==5.3.4
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.104.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.25.1
idna==3.10
Jinja2==3.1.2
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.3.0
prompt_toolkit==3.0.48
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.17.1
